+++ src/masternode/masternode-utils.cpp
@@ -22,7 +22,7 @@ void CMasternodeUtils::ProcessMasternodeConnections(CConnman& connman)
     std::vector<CDeterministicMNCPtr> vecDmns; // will be empty when no wallet
 
     // Don't disconnect masternode connections when we have less then the desired amount of outbound nodes
-    int nonMasternodeCount = 0;
+    size_t nonMasternodeCount = 0;
     connman.ForEachNode(AllNodes, [&](CNode* pnode) {
         if (!pnode->IsInboundConn() && !pnode->IsFeelerConn() && !pnode->IsManualConn() && !pnode->fMasternode && !pnode->fMasternodeProbe) {
             nonMasternodeCount++;
